arreglos generales del tema, un poco de orden
