+++ wp-content/themes/bioTienda/components/sections/_front-blog.php
@@ -1,10 +1,10 @@
 <?php
 $args = ['title' => 'Blog'];
+$blogPost = new WP_Query(['post_type' => 'post', 'post_per_page' => 4,]);
 // ----------------------------- Title
 echo get_template_part(COMPONENTS . '_title', 'large', $args) ?> 
 <section id="blog" class="grid grid-cols-1 gap-5 lg:grid-cols-4 m-ext">
     <?php
-    $blogPost = new WP_Query(['post_type' => 'post', 'post_per_page' => 4,]);
     if ($blogPost->have_posts()) :
         while ($blogPost->have_posts()) :
             $blogPost->the_post();
