+++ wp-content/themes/bioTienda/functions.php
@@ -24,7 +24,7 @@ function enqueue_scripts(){
     wp_enqueue_script('iconify', 'https://code.iconify.design/iconify-icon/1.0.0-beta.2/iconify-icon.min.js', '', '1.0', 'all');
     wp_enqueue_script('swiper', 'https://cdn.jsdelivr.net/npm/swiper@10/swiper-bundle.min.js', '', '1.8.1', 'all');
     wp_enqueue_script('bt_npm_imports', get_stylesheet_directory_uri() . '/node_modules/drift-zoom/dist/Drift.js', '', '1.8.1', 'all');
-    wp_enqueue_script('bt_npm_anime', get_stylesheet_directory_uri() . '/node_modules/animejs/lib/anime.js', '', '1.8.1', 'all');
+    wp_enqueue_script('bt_npm_anime', get_stylesheet_directory_uri() . '/node_modules/animejs/lib/anime.min.js', '', '1.8.1', 'all');
     wp_enqueue_script('bt_swiper', get_stylesheet_directory_uri() . '/assets/js/swipers.js', 'swiper', '1.8.1', 'all');
     wp_enqueue_script('bt_buttons', get_stylesheet_directory_uri() . '/assets/js/buttons.js', '', '1.0', 'all');
     wp_enqueue_script('hoverInteractions', get_stylesheet_directory_uri() . '/assets/js/hoverInteractions.js', '', '1.8.1', 'all');
