+++ wp-content/themes/bioTienda/inc/replacement_functions/load.php
@@ -1,5 +1,5 @@
 <?php
-require 'woocommerce_loop_intro.php';
+require 'woocommerce_loop.php';
 require 'woocommerce_single-product.php';
 require 'woocommerce_single-product_title.php';
 require 'archive-content.php';
